backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
python-dotenv
pydantic
passlib[bcrypt]
python-jose[cryptography]
pytest==6.2.4
pytest-asyncio==0.15.1
httpx==0.18.2
pytest-mock==3.6.1
aiomysql
testcontainers
pytest-cov